+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
-Upcoming bug-fix release, possbily as early as 1 August 2026.
+Upcoming feature release, possbily as early as 1 August 2026.
 
 ### Fixed
 
@@ -19,11 +19,21 @@ Upcoming bug-fix release, possbily as early as 1 August 2026.
 
 ### Added
 
+ - #32: Added `XBoolean` as the xchange-specific boolean type, replacing `boolean` in prior versions. (The `boolean`
+   type is still defined for compatibility, but can be disabled by defining `_TYPEDEF_BOOLEAN` prior to including
+   `xchange.h` in your source code.).
+
  - Added `xIsDebug()` function to check on `xDebug`. While the global variable is fine in most cases, they are 
    problematic for Windows DLLs. It's better to use purely functional access instead.
    
  - `xvprintf()` / `xdprintf()` macros to reference functions instead of global vars (see above comment on Windows
    DLLs.
+   
+### Changed
+
+ - #32: Changed `boolean` parameter and return types to the equivalent `XBoolean`, to disambugiate the 
+   xchange-specific boolean type from other definitions. (Other header may define `boolean` also, and so by choosing
+   a more unique type name, we reduce the chance of namespace conflicts.)
    
 
 ## [1.2.0] - 2026-06-08